# nfactorial-frontend/src/shared/api/generated/model/character.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Rick and Morty API
 * The Rick and Morty API documentation
 * OpenAPI spec version: 1.0
 */
import type { CharacterOrigin } from './characterOrigin';
import type { CharacterLocation } from './characterLocation';

export interface Character {
  /** The ID of the character */
  id: number;
  /** The name of the character */
  name: string;
  /** The status of the character (alive, dead or unknown) */
  status: string;
  /** The species of the character */
  species: string;
  /** The type or subspecies of the character */
  type: string;
  /** The gender of the character (female, male, genderless or unknown) */
  gender: string;
  /** The character's origin location */
  origin: CharacterOrigin;
  /** The character's last known location */
  location: CharacterLocation;
  /** Link to the character's image */
  image: string;
  /** List of episodes in which this character appeared */
  episode: string[];
  /** Link to the character's own URL endpoint */
  url: string;
  /** Time at which the character was created in the database */
  created: string;
}