# nfactorial-frontend/src/shared/api/generated/model/info.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Rick and Morty API
 * The Rick and Morty API documentation
 * OpenAPI spec version: 1.0
 */

export interface Info {
  /** The length of the response */
  count: number;
  /** The amount of pages */
  pages: number;
  /** Link to the next page (if it exists) */
  next: string;
  /** Link to the previous page (if it exists) */
  prev: string;
}