# nfactorial-frontend/src/shared/api/generated/model/location.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Rick and Morty API
 * The Rick and Morty API documentation
 * OpenAPI spec version: 1.0
 */

export interface Location {
  /** The ID of the location */
  id: number;
  /** The name of the location */
  name: string;
  /** The type of the location */
  type: string;
  /** The dimension in which the location is located */
  dimension: string;
  /** List of characters who have been last seen in the location */
  residents: string[];
  /** Link to the location's own URL endpoint */
  url: string;
  /** Time at which the location was created in the database */
  created: string;
}